+++ README.md
@@ -1,10 +1,19 @@
 # libsm64-panda
 (still super heavy WIP) Panda3D Python bindings for libsm64
 
-# How?
+## How?
 Check out libsm64: https://github.com/libsm64/libsm64
 
-# Running
+## Running
 You'll need to build libsm64 yourself - as well as have a copy of the Super Mario 64 US rom  
 Put both (named "sm64" with respective file extensions) in the root project directory  
-And to run this, open your favorite command line and run ``ppython main.py`` to run the example program  
+And to run this, open your favorite command line and run ``ppython main.py`` to run the example program .
+
+## License
+All original code in this project is licensed under either the [Apache License 2.0](./LICENSE-APACHE) or
+the [MIT License](./LICENSE-MIT), at your option. Unless you explicitly state otherwise, any contribution
+intentionally submitted for inclusion in this project by you (as defined in the Apache 2.0 license)
+shall be licensed as such, without any additional terms or conditions.
+
+Currently, the only exception to the above is [from_blender.py](./from_blender.py),
+the license details for which are still being hashed out.
